- make sure lowercase script uses Moses and not awk         - notably there was no real difference         - however, doing this for uniformity's sake - retire `scripts/lowercase-english-side.sh` - add help to scripts/segment-sentences-morfessor.py - make sure zip file is removed after downloading in   download-morfessor-models.sh

+++ scripts/lowercase.sh
@@ -9,11 +9,18 @@ if [ -z "${TRAIN_FOLDER}" ]; then
     exit 1
 fi
 
+if [ -z "${MOSES_SCRIPTS}" ]; then
+    echo "Please make sure the MOSES_SCRIPTS environment variable is set!"
+    exit 1
+fi
+
+LOWERCASE_SCRIPT="$MOSES_SCRIPTS/tokenizer/lowercase.perl"
+
 cd "${TRAIN_FOLDER}" || exit
 
 for input_file in ./*; do
     output_file=$input_file"_new"
-    awk '{ print tolower($0) }' \
+    $LOWERCASE_SCRIPT \
         <"${input_file}" \
         >"${output_file}"
     rm "${input_file}"
